Prune pipelines
